Adapt the value of max_depth if needed
Although the generated fuzzer is aware of the depth of the minimal
derivation of a certain rule and hence it's able to ensure finishing
the generation in a predefined depth, however this only works if there
are no inline predicates in the grammar. In such cases, it can navigate
into unresolvable situations where the available derivatations are
disabled by the predicates. The current patch solves such situations
by temporarily raising the max depth to the minimal required value and
it will finish the generation with choosing the shortest possible
derivation path.